PUD: replace rt.jar file with ee.j2se.jar
We now use an API-only jar file which is much smaller and
has a more permissive license (APL 2.0) than the OpenJDK
rt.jar file (GPLv2 with classpath exception).

Note that this jar file is only used at build time, it
is _not_ included in any of the binaries. Please refer
to the lib/pud/wireformat-java/README for an explantion.

Signed-off-by: Ferry Huberts <ferry.huberts@pelagic.nl>