mdns: do not close an fd that is not open
Coverity:
CID 739659 (#1 of 1): Argument cannot be negative (NEGATIVE_RETURNS)
At (7): "capturingSkfd" is passed to a parameter that cannot be negative.
CID 739660 (#1 of 1): Argument cannot be negative (NEGATIVE_RETURNS)
At (8): "electionSkfd" is passed to a parameter that cannot be negative.
CID 739661 (#1 of 1): Argument cannot be negative (NEGATIVE_RETURNS)
At (9): "helloSkfd" is passed to a parameter that cannot be negative.

CID 739656 (#6 of 14): Argument cannot be negative (NEGATIVE_RETURNS)
At (7): "capturingSkfd" is passed to a parameter that cannot be negative.
CID 739656 (#9 of 14): Argument cannot be negative (NEGATIVE_RETURNS)
At (7): "electionSkfd" is passed to a parameter that cannot be negative.
CID 739656 (#10 of 14): Argument cannot be negative (NEGATIVE_RETURNS)
At (7): "helloSkfd" is passed to a parameter that cannot be negative.

Signed-off-by: Ferry Huberts <ferry.huberts@pelagic.nl>