Fix setsockopt for setting Traffic Class in IPv6
HiOLSR-developers,

For your consideration:

Setting Traffic Class (the counterpart of TOS in IPv4) on OLSR IPv6 packets is currently broken.
The code does not distinguish between the IPv4 and IPv6 cases, but it should, as the required
'level' and 'optname' arguments to setsockopt differ.

I tested this patch by monitoring OLSR IPv6 packets with 'tcpdump -v': it appears to be working.

Note this is for Linux only.

This is my modest contribution to Dave Täht's quest to fix "TOS setting" in IPv6 :-)
See: http://www.bufferbloat.net/issues/249

Thanks,
Ronald

Signed-off-by: Ronald in 't Velt <velt@xs4all.nl>