pud: remove a strlen in the DNS name handling of the wire format.
This makes the handling of DNS names in the wire format much more robust.

This commit prepares for not having a terminating '\0' for a DNS name in
the wire format. However, since that is a breaking change, we don't
actually make that change until a later point in time at which we're
forced to break compatibility anyway.

The benefit of saving a single byte in this case is too minor to justify
breaking the compatibility.

Signed-off-by: Ferry Huberts <ferry.huberts@pelagic.nl>