gw: move path_cost into the gateway struct, ensure it's always up-to-date
In the process a lot of code surrounding the path costs can be simplified:
upon reception of a HNA we always calculate the costs, and when they have
changed, we re-sort the gateway list.
Therefore no other place in the code needs to do that, and path_cost can
always be used directly from the gateway struct.

Signed-off-by: Ferry Huberts <ferry.huberts@pelagic.nl>