jsoninfo: HTTP headers with CORS (if requested)
The new "httpheaders" parameter prepends HTTP headers to the reply.
If not set it will default to "no" and have the same behaviour as before.
Cross-origin resource sharing headers (CORS) are included in reply allowing the
json retrieval by javascript applications not served by olsrd itself.
This will allow to easily develop js applications running directly in the
browser.

Reviewed-by: Ferry Huberts <ferry.huberts@pelagic.nl>