sgw: send sgw HNAs with zero bandwidth after local gw has gone away
This is to improve sgw convergence speed in the mesh.

sgw HNAs with zero bandwidth are sent after the local default gateway
has gone away (it's bandwidth has become zero on either the uplink,
downlink or both) or its interface has gone down or disappeared.

These HNAs are only sent during the next 'HNA period' after which the
sending of these HNAs stops; this is the 'unannounce' window.

Each of the zero bandwidth HNAs has an adjusted vtime so that the actual
timeout of the original HNA on the receiver side doesn't change (within
the bound of the normal jittering of sending the HNAs).

Signed-off-by: Ferry Huberts <ferry.huberts@pelagic.nl>