sgw: do not send zero bandwidth sgw HNAs by default
Previously, setting the uplink and/or downlink speed to zero would
cause zero bandwidth HNAs to be sent.

This patch ensures that these HNAs will not be sent, only after the
bandwidth was non-zero and it has become zero will these HNAs be sent
(so only in the 'unannounce' window of the previous patch).

Signed-off-by: Ferry Huberts <ferry.huberts@pelagic.nl>